extract validation
