+++ lib/ws_servers/api/algos/algo_worker.js
@@ -3,16 +3,10 @@
 const debug = require('debug')
 const d = debug('bfx:hf:server:algo-worker')
 const { AOHost } = require('bfx-hf-algo')
-const _get = require('lodash/get')
 
 const validateAO = require('./util/validate_ao')
+const { hasRequiredPermissions } = require('./util/validate_permissions')
 const { DMS_ENABLED } = require('../../../constants')
-
-const requiredPermissions = [
-  'orders.read',
-  'orders.write',
-  'wallets.read'
-]
 
 class AlgoWorker {
   constructor (settings, algoOrders, bcast, algoDB, logAlgoOpts, marketData, config = {}) {
@@ -72,7 +66,7 @@ class AlgoWorker {
     this.registerEvents()
 
     const { caps } = await host.connect()
-    if (!this.hasRequiredPermissions(caps)) {
+    if (!hasRequiredPermissions(caps)) {
       throw new Error('The given API key does not have the required permissions, please make sure to enable "Get orders and statuses", "Create and cancel orders", "Get wallet balances and addresses"')
     }
 
@@ -320,10 +314,6 @@ class AlgoWorker {
 
       throw new Error('Failed to start algo order')
     }
-  }
-
-  hasRequiredPermissions (caps) {
-    return requiredPermissions.every(perm => _get(caps, perm))
   }
 }
